change Path to use Core2
